+++ Milestone 1/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: abnsila <abnsila@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/05 16:48:20 by abnsila           #+#    #+#             */
-/*   Updated: 2024/12/06 18:21:15 by abnsila          ###   ########.fr       */
+/*   Updated: 2024/12/07 14:41:43 by abnsila          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_printf(const char *format, ...);
 int ft_print_char(int c);
 int ft_print_str(char *s);
 int ft_print_nbr(int n);
-int ft_print_hexa(unsigned long int n, char *base);
+int ft_print_hexa(unsigned int n, char *base);
 int ft_print_ptr(void *p);
 int ft_print_unsigned_nbr(unsigned int n);
